requirements.txt
joblib==0.15.1
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
Flask==1.1.2
numpy==1.18.4
plac==1.1.3
Flask_Cors==3.0.8
flask_bcrypt==0.7.1
flask_jwt_extended==3.24.1
flask_mysqldb==0.2.0
flask_restful==0.3.8
gunicorn==20.0.4
keras==2.3.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
mysql_connector_repackaged==0.3.1
nltk==3.5
pandas==1.0.4
PyJWT==1.7.1
python-dotenv==0.13.0
tensorflow==1.14.0
